cleanup,test[#70]: S3Service 코드 구조 및 가독성 개선하기

+++ src/main/java/com/ktb/howard/ktb_community_server/image/domain/Image.java
@@ -61,7 +61,7 @@ public class Image extends BaseEntity {
     private String fileName;
 
     @Column(name = "file_size", nullable = false)
-    private Integer fileSize;
+    private Long fileSize;
 
     @Column(name = "mime_type", nullable = false)
     private String mimeType;
@@ -82,7 +82,7 @@ public class Image extends BaseEntity {
                  String region,
                  String objectKey,
                  String fileName,
-                 Integer fileSize,
+                 Long fileSize,
                  String mimeType,
                  Integer sequence,
                  ImageStatus status
