+++ src/main/java/com/ktb/howard/ktb_community_server/image/dto/CreateImageUploadUrlRequestDto.java
@@ -19,7 +19,7 @@ public class CreateImageUploadUrlRequestDto {
     public record ImageUploadRequestInfoDto(
             ImageType imageType,
             String fileName,
-            Integer fileSize,
+            Long fileSize,
             String mimeType,
             Integer sequence
     ) { }
